Move local_now() to utilities.datetime
